# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/oauth/applications/edit": {
      "put": {
        "operationId": "OAuthProviderController_editUserApplication",
        "summary": "Edit an existing OAuth client application",
        "parameters": [
          {
            "name": "application_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OAuth Provider"
        ]
      }
    },
    "/oauth/authorize": {
      "get": {
        "operationId": "OAuthProviderController_authorize",
        "summary": "Authorize an OAuth client",
        "parameters": [
          {
            "name": "client_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "redirect_uri",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "response_type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OAuth Provider"
        ]
      }
    },
...
    "/oauth/consent": {
      "get": {
        "operationId": "OAuthProviderController_getConsentScreen",
        "summary": "Get consent screen information",
        "parameters": [
          {
            "name": "client_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OAuth Provider"
        ]
      },
      "post": {
        "operationId": "OAuthProviderController_grantConsent",
        "summary": "Grant consent for an application",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "OAuth Provider"
        ]
      }
    },
...
    "/v1/user/change-fullName": {
      "patch": {
        "operationId": "UserController_changeFullName",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeNameDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/v1/user/change-birthdate": {
      "patch": {
        "operationId": "UserController_changeBirthDate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeBirthdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/v1/user/change-profilePicture": {
      "patch": {
        "operationId": "UserController_changeProfilePicture",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePhotoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/v1/collection/create": {
      "post": {
        "operationId": "CollectionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlbumDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
    "/v1/collection/update": {
      "put": {
        "operationId": "CollectionController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbumDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
    "/v1/collection/delete": {
      "delete": {
        "operationId": "CollectionController_delete",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAlbumDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Collections"
        ]
      }
    },
...
  },
  "info": {
    "title": "XENAuth API",
    "description": "The official XENAuth API documentation",
    "version": "1.0",
    "contact": {
      "name": "Erzen Krasniqi",
      "url": "https://erzen.tk",
      "email": "erzenkrasniqi@matrics.io"
    },
    "termsOfService": "https://erzen.tk/terms",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    }
  },
...
  "servers": [
    {
      "url": "https://localhost:3000",
      "description": "LOCAL server"
    },
    {
      "url": "https://api.erzen.xyz",
      "description": "DEV server"
    },
    {
      "url": "https://apis.erzen.xyz",
      "description": "PRODUCTION server"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChangeNameDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "ChangeBirthdateDto": {
        "type": "object",
        "properties": {
          "birthdate": {
            "type": "string",
            "description": "User new birthdate"
          }
        },
        "required": [
          "birthdate"
        ]
      },
      "ChangePhotoDto": {
        "type": "object",
        "properties": {
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "User new photo"
          }
        },
        "required": [
          "photo"
        ]
      },
      "CreateAlbumDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the album"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateAlbumDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the album"
          },
          "id": {
            "type": "number",
            "description": "The ID of the album"
          }
        },
        "required": [
          "title",
          "id"
        ]
      },
      "DeleteAlbumDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "The ID of the album to delete"
          }
        },
        "required": [
          "id"
        ]
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}